Move Gilda client code into grounding_mapper module
