+++ indra/tests/test_groundingmapper.py
@@ -1,6 +1,7 @@
 from indra.preassembler.grounding_mapper import default_mapper as gm
 from indra.preassembler.grounding_mapper import GroundingMapper
 from indra.preassembler.grounding_mapper.analysis import *
+from indra.preassembler.grounding_mapper.gilda import ground_statements
 from indra.statements import Agent, Phosphorylation, Complex, Inhibition, \
     Evidence, BoundCondition
 from indra.util import unicode_strs
@@ -395,3 +396,15 @@ def test_misgrounding():
     assert len(stmt.sub.db_refs) == 1, stmt.sub.db_refs
     assert stmt.sub.db_refs['TEXT'] == 'baz1'
     assert stmt.sub.name == 'baz1'
+
+
+def test_ground_gilda():
+    mek = Agent('Mek', db_refs={'TEXT': 'MEK'})
+    erk = Agent('Erk1', db_refs={'TEXT': 'Erk1'})
+    stmt = Phosphorylation(mek, erk)
+    ground_statements([stmt])
+    assert stmt.enz.name == 'MEK', stmt.enz
+    assert stmt.enz.db_refs['FPLX'] == 'MEK'
+    assert stmt.sub.name == 'MAPK3'
+    assert stmt.sub.db_refs['HGNC'] == '6877'
+    assert stmt.sub.db_refs['UP'] == 'P27361'
